I've faced issue "LBA out of range"
Bochs will start esi register with 0x000e0000 that's why we have to clear this. Because we uses esi register as sector index in load_sector. We've changed low byte of esi register but high byte is remain as 0x000e. After incl %esi instruction for sector 1, we will get 0x000e0001 that's why Bochs gives the "LBA out of range".